feature: update loading

+++ lib/models/responses/home_detail_response.dart
@@ -85,7 +85,7 @@ class Data {
   int? view;
   bool? favorite;
   String? thumbnail;
-  int? averageRate;
+  double? averageRate;
   int? numberOfReviews;
   DateTime? lastModifiedDate;
   List<RoomsImportant>? roomsImportant;
